requirements.txt
Django==4.2.3
django-environ
environ==1.0
mysqlclient==2.1.0
google-news-feed==1.1.0
news-please
newspaper3k
django_crontab
scikit-learn
djangorestframework
PyJWT
django-cors-headers